Move CompactVectorOfStrings to its own header util/CompactStringVector.h
CompactVectorOfStrings (and its detail::CompactStringVectorWriter helper)
is unrelated to the Pattern struct that gives global/Pattern.h its name;
give it its own header and fix up all direct/transitive includes:

- CountAvailablePredicates.h, HasPredicateScan.h, IndexImpl.h,
  VocabularyInMemory.h, VocabularyInMemoryBinSearch.h: only needed
  CompactVectorOfStrings, so their global/Pattern.h include is replaced.
- PatternCreator.h: needs both Pattern and CompactVectorOfStrings, so it
  keeps global/Pattern.h and gains util/CompactStringVector.h.
- CountAvailablePredicates.cpp, HasPredicateScanTest.cpp,
  LocalVocabTest.cpp: used Pattern::NoPattern transitively through a
  header that no longer provides it; add the direct include.
