+++ src/index/IndexImpl.h
@@ -17,7 +17,6 @@
 #include "backports/algorithm.h"
 #include "engine/Result.h"
 #include "engine/idTable/CompressedExternalIdTable.h"
-#include "global/Pattern.h"
 #include "global/SpecialIds.h"
 #include "index/CompressedRelation.h"
 #include "index/ConstantsIndexBuilding.h"
@@ -38,6 +37,7 @@
 #include "parser/RdfParser.h"
 #include "parser/TripleComponent.h"
 #include "util/BufferedVector.h"
+#include "util/CompactStringVector.h"
 #include "util/File.h"
 #include "util/Forward.h"
 #include "util/Iterators.h"
