+++ src/index/PatternCreator.h
@@ -16,6 +16,7 @@
 #include "index/ConstantsIndexBuilding.h"
 #include "index/ExternalSortFunctors.h"
 #include "util/BufferedVector.h"
+#include "util/CompactStringVector.h"
 #include "util/ExceptionHandling.h"
 #include "util/HashMap.h"
 #include "util/Serializer/Serializer.h"
