+++ src/index/vocabulary/VocabularyInMemoryBinSearch.h
@@ -8,10 +8,10 @@
 #include <string>
 #include <string_view>
 
-#include "global/Pattern.h"
 #include "index/vocabulary/VocabularyBinarySearchMixin.h"
 #include "index/vocabulary/VocabularyTypes.h"
 #include "util/Algorithm.h"
+#include "util/CompactStringVector.h"
 #include "util/Exception.h"
 #include "util/Serializer/FileSerializer.h"
 #include "util/Serializer/SerializeVector.h"
